requirements.txt
flask==2.3.3
easyocr==1.7.0
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3